Added php8 support (release 3.1.0)
This commit adds PHP8 support to the contentful core. It does the
following:
- Bump the version to 3.1.0 (new minor)
- Add a php8 image to the CI to run the tests on php8
- Allow PHP8 in composer.json

The code was verified against the official breaking changes guide [0].
Note that code coverage is not yet supported by phpunit on php8. As it
should be pretty much the same as on older PHP versions, however, this
is accepted for now.

[0] https://www.php.net/manual/en/migration80.php